Refactor some of the register/views.py, moved less readable code into register/help folder and broke it apart to help readability
Also cleaned up some of the imports and removed a bunch of code altogether.

Tests ran fine, but I want more ais2 testing than usual to make sure I didnt miss something
