+++ register/views.py
@@ -1,56 +1,19 @@
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse
 from django.utils import timezone
-from django.utils.timezone import utc
 from django.contrib.auth import authenticate, login, update_session_auth_hash
 from django.conf import settings
-from django.core.mail import send_mail
-from django.template.loader import get_template
-from django.contrib.sites.shortcuts import get_current_site
-
-from collections import namedtuple
-import math
-
-import json
-import requests as r
-
-import datetime
 
 from companies.models import Company, Contact
-from orders.models import Product, Order, ProductType
+from orders.models import Product, Order
 from exhibitors.models import Exhibitor
 from fair.models import Fair
 from sales.models import Sale
-from matching.models import Survey, Question, Response, TextAns, ChoiceAns, IntegerAns, BooleanAns
-from .models import SignupContract, SignupLog, OrderLog
-from .forms import CompanyForm, ContactForm, RegistrationForm, CreateContactForm, UserForm, InterestForm, ExhibitorForm, ChangePasswordForm
 
-BASE_PRICE = 39500
-PRODUCT_LOG = ":"
+from .models import SignupContract, SignupLog
+from .forms import CompanyForm, ContactForm, RegistrationForm, CreateContactForm, UserForm, InterestForm, ChangePasswordForm
 
-def getTimeFlag(close_offset = 7, warning_offset = 7):
-    # used to close cr, a warning text after deadline will pop up, however exhibitors will not be permitted to do any changes after the offset in days has passed
-    currentFair = None
-    try:
-        currentFair = Fair.objects.get(current=True)
-        if currentFair.complete_registration_close_date:
-            end_time = currentFair.complete_registration_close_date.replace(tzinfo=utc)
-            end_time_close = end_time + datetime.timedelta(days=close_offset)
-            time = datetime.datetime.now().replace(tzinfo=utc)
-            time = time.replace(microsecond=0)
-            warning_time = end_time - datetime.timedelta(days=warning_offset)
-            if time < end_time and time > warning_time:
-                return('warning', [end_time, end_time - time])
-            elif time > end_time and time < end_time_close:
-                return('overdue', [end_time, time - end_time])
-            elif time > end_time_close:
-                return('closed', [end_time, time - end_time])
-            else:
-                return(None, [None, None])
-        else:
-            return(None, [None, None])
-    except Fair.DoesNotExist:
-        return(None, [None, None])
+from .help import exhibitor_form as help
+from .help.methods import get_time_flag
 
 def index(request, template_name='register/index.html'):
     if request.user.is_authenticated():
@@ -58,7 +21,7 @@ def index(request, template_name='register/index.html'):
             return redirect('anmalan:home')
         else:
             return redirect('anmalan:logout')
-    timeFlag, [time_end, time_diff] = getTimeFlag()
+    timeFlag, [time_end, time_diff] = get_time_flag()
     return render(request, template_name, {'timeFlag': timeFlag, 'time_end': time_end, 'time_diff': time_diff})
 
 def home(request, template_name='register/home.html'):
@@ -200,7 +163,6 @@ def create_exhibitor(request, template_name='register/exhibitor_form.html'):
 
      """
 
-    productLog = "Base kit \n"
     currentFair = Fair.objects.get(current = True)
     # Return 404 if no contract.
     # if no contact or company connected to
@@ -224,390 +186,19 @@ def create_exhibitor(request, template_name='register/exhibitor_form.html'):
             except Exhibitor.DoesNotExist:
                 pass
 
-            # Get products which requires an amount and put them into the Exhibitor form
-            banquet_products = Product.objects.filter(fair=currentFair, product_type=ProductType.objects.filter(name="Banquet"))
-            lunch_products = Product.objects.filter(fair=currentFair, product_type=ProductType.objects.filter(name="AdditionalLunch"))
-            event_products = Product.objects.filter(fair=currentFair, product_type=ProductType.objects.filter(name="Events"))
-            room_products = Product.objects.filter(fair=currentFair, product_type=ProductType.objects.filter(name="Rooms"))
-            nova_products = Product.objects.filter(fair=currentFair, product_type=ProductType.objects.filter(name="Nova"))
-            stand_area_products = Product.objects.filter(fair=currentFair, product_type=ProductType.objects.filter(name="Additional Stand Area"))
-            stand_height_products = Product.objects.filter(fair=currentFair, product_type=ProductType.objects.filter(name="Additional Stand Height"))
-
-            # Check which products that already is in an order
-            current_banquet_orders = Order.objects.filter(exhibitor=exhibitor, product__in=banquet_products)
-            current_lunch_orders = Order.objects.filter(exhibitor=exhibitor, product__in=lunch_products)
-            current_event_orders = Order.objects.filter(exhibitor=exhibitor, product__in=event_products)
-            current_room_orders = Order.objects.filter(exhibitor=exhibitor, product__in=room_products)
-            current_nova_orders = Order.objects.filter(exhibitor=exhibitor, product__in=nova_products)
-            current_stand_area_orders = Order.objects.filter(exhibitor=exhibitor, product__in=stand_area_products)
-            current_stand_height_orders = Order.objects.filter(exhibitor=exhibitor, product__in=stand_height_products)
-
-            # get survey and corresponding matching questions
-            try:
-                matching_survey = Survey.objects.get(fair=currentFair, name='exhibitor-matching')
-            except Survey.DoesNotExist:
-                matching_survey = None
-            matching_questions = Question.objects.filter(survey=matching_survey)
-            # check which questions are already answered
-            current_matching_responses = Response.objects.filter(exhibitor=exhibitor, survey=matching_survey)
-
-            # Set automated time closing of cr
-            timeFlag, time_disp = getTimeFlag()
-
-            # Pass along all relevant information to form
-            form = ExhibitorForm(
-                request.POST or None,
-                request.FILES or None,
-                instance = exhibitor,
-                banquet = banquet_products,
-                lunch = lunch_products,
-                events = event_products,
-                rooms = room_products,
-                nova = nova_products,
-                stand_area = stand_area_products,
-                stand_height = stand_height_products,
-                banquet_orders = current_banquet_orders,
-                lunch_orders = current_lunch_orders,
-                event_orders = current_event_orders,
-                room_orders = current_room_orders,
-                nova_orders = current_nova_orders,
-                stand_area_orders = current_stand_area_orders,
-                stand_height_orders = current_stand_height_orders,
-                company = company,
-                contact = contact,
-                matching_survey = matching_survey,
-                matching_questions = matching_questions,
-                matching_responses = current_matching_responses,
-                timeFlag = timeFlag,
-                time_disp = time_disp,
-            )
+            form = help.create_exhibitor_form(request, currentFair, exhibitor, company, contact)
 
             if form.is_valid():
-                # get selected products. IMPORTANT: NEEDS TO BE BEFORE form.save(commit=False)
-                product_selection_rooms = form.cleaned_data['product_selection_rooms']
-                product_selection_nova = form.cleaned_data['product_selection_nova']
-                product_selection_additional_stand_area = form.cleaned_data['product_selection_additional_stand_area']
-                product_selection_additional_stand_height = form.cleaned_data['product_selection_additional_stand_height']
-
-                # Save exhibitor model values from form into exhibitor variable
-                exhibitor = form.save(commit=False)
-
-                # Update Company fields
-                updatedCompany = Company.objects.get(pk=company.pk)
-                updatedCompany.organisation_number = form.cleaned_data['organisation_identification_number']
-                updatedCompany.organisation_type = form.cleaned_data['type_of_organisation']
-                updatedCompany.address_street = form.cleaned_data['address_street']
-                updatedCompany.address_zip_code = form.cleaned_data['address_zip_code']
-                updatedCompany.address_city = form.cleaned_data['address_city']
-                updatedCompany.address_country = form.cleaned_data['address_country']
-                updatedCompany.additional_address_information = form.cleaned_data['additional_address_information']
-                updatedCompany.website = form.cleaned_data['website']
-                updatedCompany.save()
-                exhibitor.company = updatedCompany
-
-                # Update Contact fields
-                updatedContact = Contact.objects.get(pk=contact.pk)
-                updatedContact.name = form.cleaned_data['contact_name']
-                updatedContact.work_phone = form.cleaned_data['work_phone']
-                updatedContact.cell_phone = form.cleaned_data['cell_phone']
-                updatedContact.phone_switchboard = form.cleaned_data['phone_switchboard']
-                updatedContact.email = form.cleaned_data['contact_email']
-                updatedContact.alternative_email = form.cleaned_data['alternative_email']
-                updatedContact.save()
-                exhibitor.contact = updatedContact
-
-                # other exhibitor fields that you do not choose in the form
-                exhibitor.fair = currentFair
-
-                # Create or update exhibitor
-                try:
-                    exhibitor.pk = Exhibitor.objects.get(company=exhibitor.company, fair=currentFair).pk
-                    exhibitor.save()
-                except Exhibitor.DoesNotExist:
-                    exhibitor.save()
-
-                # create or update orders to the current exhibitor from products
-                def create_or_update_order(product, amount):
-                        order = None
-                        try:
-                            order = Order.objects.get(product=product,exhibitor=exhibitor)
-                            order.amount = amount
-                            order.save()
-                        except Order.DoesNotExist:
-                            order = Order.objects.create(
-                                exhibitor=exhibitor,
-                                product=product,
-                                amount=amount,
-                            )
-                # delete an order via a product and the current exhibitor
-                def delete_order_if_exists(product):
-                    try:
-                        Order.objects.get(product=product, exhibitor=exhibitor).delete()
-                    except Order.DoesNotExist:
-                        return
-
-                def product_amount_string(product, amount):
-                    return product.name + " x " + str(amount) + "\n"
-
-                # Create or update orders from the checkbox products (ProductMultiChoiceField).
-                # If they are not checked in but exist as an order in db, then delete.
-                room_products = Product.objects.filter(fair=Fair.objects.get(current = True), product_type=ProductType.objects.filter(name="Rooms"))
-                nova_products = Product.objects.filter(fair=Fair.objects.get(current = True), product_type=ProductType.objects.filter(name="Nova"))
-                stand_area_products = Product.objects.filter(fair=Fair.objects.get(current = True), product_type=ProductType.objects.filter(name="Additional Stand Area"))
-                stand_height_products = Product.objects.filter(fair=Fair.objects.get(current = True), product_type=ProductType.objects.filter(name="Additional Stand Height"))
-
-                # Boolean (checkmark) products
-                bool_products = []
-				# TODO fetch prices from elsewhere
-                total_price = BASE_PRICE
-
-
-                for product in room_products:
-                    if product in product_selection_rooms:
-                        bool_products.append(product)
-                        total_price += product.price
-                        create_or_update_order(product, 1)
-                        productLog += product_amount_string(product, 1)
-                    else:
-                        delete_order_if_exists(product)
-                for product in nova_products:
-                    if product in product_selection_nova:
-                        bool_products.append(product)
-                        total_price += product.price
-                        create_or_update_order(product, 1)
-                        productLog += product_amount_string(product, 1)
-                    else:
-                        delete_order_if_exists(product)
-                for product in stand_area_products:
-                    # this is a fix due to replace in products_as_select_field foo in register/forms.py, which is needed for the js that generates a product list in the confirm and submit tab
-                    option = str(product.name)
-                    option = option.replace(" ", "")
-                    option = option.replace(",", "_")
-                    if option in product_selection_additional_stand_area:
-                        bool_products.append(product)
-                        total_price += product.price
-                        create_or_update_order(product, 1)
-                        productLog += product_amount_string(product, 1)
-                    else:
-                        delete_order_if_exists(product)
-                for product in stand_height_products:
-                    # this is a fix due to replace in products_as_select_field foo in register/forms.py, which is needed for the js that generates a product list in the confirm and submit tab
-                    option = str(product.name)
-                    option = option.replace(" ", "")
-                    option = option.replace(",", "_")
-                    if option in product_selection_additional_stand_height:
-                        bool_products.append(product)
-                        total_price += product.price
-                        create_or_update_order(product, 1)
-                        productLog += product_amount_string(product, 1)
-                    else:
-                        delete_order_if_exists(product)
-
-				# Numerical (amount) products
-                num_products = []
-                NumProduct = namedtuple('NumProduct', ['name', 'amount', 'price'])
-
-                # Create or update orders from products that can be chosen in numbers.
-                # If they have an amount equal to zero then delete the order.
-                # Try if amount is None
-                for (banquetProduct, amount) in form.amount_products('banquet_'):
-                    try:
-                        if amount > 0:
-                            create_or_update_order(banquetProduct, amount)
-                            productLog += product_amount_string(banquetProduct, amount)
-                            num_products.append(NumProduct(banquetProduct.name, amount, amount * banquetProduct.price))
-                            total_price += amount * banquetProduct.price
-                        else:
-                            delete_order_if_exists(banquetProduct)
-                    except TypeError:
-                        amount = 0
-                        delete_order_if_exists(banquetProduct)
-
-                for (lunchProduct, amount) in form.amount_products('lunch_'):
-                    try:
-                        if amount > 0:
-                            create_or_update_order(lunchProduct, amount)
-                            productLog += product_amount_string(lunchProduct, amount)
-                            num_products.append(NumProduct(lunchProduct.name, amount, amount * lunchProduct.price))
-                            total_price += amount * lunchProduct.price
-                        else:
-                            delete_order_if_exists(lunchProduct)
-                    except TypeError:
-                        amount = 0
-                        delete_order_if_exists(lunchProduct)
-
-                for (eventProduct, amount) in form.amount_products('event_'):
-                    try:
-                        if amount > 0:
-                            create_or_update_order(eventProduct, amount)
-                            productLog += product_amount_string(eventProduct, amount)
-                            num_products.append(NumProduct(eventProduct.name, amount, amount * eventProduct.price))
-                            total_price += amount * eventProduct.price
-                        else:
-                            delete_order_if_exists(eventProduct)
-
-                    except TypeError:
-                        amount = 0
-                        delete_order_if_exists(eventProduct)
-
-
-				# Longest name length for padding purposes
-                def getNameLen(item):
-                    return len(item.name)
-                def getAmount(item):
-                    return item.amount
-
-                max_name_len_bool = 0
-                max_name_len_num = 0
-                max_amount = 0
-
-                if bool_products:
-                    max_name_len_bool = len(max(bool_products, key=getNameLen).name)
-
-                if num_products:
-                    max_name_len_num = len(max(num_products, key=getNameLen).name)
-                    max_amount = math.ceil(max(num_products, key=getAmount).amount / 10)
-
-                max_name_len = max(max_name_len_bool, max_name_len_num)
-
-
-
-                # Add the Base Kit (mandatory) and Banquet ticket - Base Kit (2 are included)
-                # If already added, don't add to db, else add.
-                base_kit_products = Product.objects.filter(fair=currentFair, product_type=ProductType.objects.filter(name="Base Kit"))
-                current_base_kit_orders = Order.objects.filter(exhibitor=exhibitor, product__in=base_kit_products)
-
-                for product in base_kit_products:
-                    if product.name == "Banquet Ticket - Base Kit":  # 2 Banquet tickets are included
-                        amount = 2
-                    else:
-                        amount = 1
-
-                    if not product in current_base_kit_orders:
-                        create_or_update_order(product, amount)
-
-                def create_or_update_answer(response, question, ans):
-                    answer = None
-                    if question.question_type == Question.TEXT:
-                        try:
-                            answer = TextAns.objects.get(question=question, response=response)
-                            answer.ans = ans
-                            answer.save()
-                        except TextAns.DoesNotExist:
-                            answer = TextAns.objects.create(question=question, response = response, ans=ans)
-                    elif question.question_type == Question.INT:
-                        try:
-                            answer = IntegerAns.objects.get(question=question, response=response)
-                            answer.ans = ans
-                            answer.save()
-                        except IntegerAns.DoesNotExist:
-                            answer = IntegerAns.objects.create(question=question, response = response, ans=ans)
-                    elif question.question_type == Question.SELECT:
-                        try:
-                            answer = ChoiceAns.objects.get(question=question, response=response)
-                            answer.ans = ans
-                            answer.save()
-                        except ChoiceAns.DoesNotExist:
-                            answer = ChoiceAns.objects.create(question=question, response = response, ans=ans)
-                    elif question.question_type == Question.BOOL:
-                        try:
-                            answer = BooleanAns.objects.get(question=question, response=response)
-                            answer.ans = ans
-                            answer.save()
-                        except BooleanAns.DoesNotExist:
-                            answer = BooleanAns.objects.create(question=question, response = response, ans=ans)
-
-
-                # create or update responses on matching questions
-                def create_or_update_response(question, ans):
-                    response = None
-                    try:
-                        response = Response.objects.get(exhibitor=exhibitor, question=question, survey=matching_survey)
-                        #response.save()
-                    except Response.DoesNotExist:
-                        response = Response.objects.create(exhibitor=exhibitor, survey=matching_survey, question=question)
-                    create_or_update_answer(response, question, ans)
-
-
-                #delete response via question and current exhibitor
-                def delete_response_if_exists(question, ans):
-                    try:
-                        Response.objects.get(exhibitor=exhibitor, survey=matching_survey, question=question).delete()
-                    except Response.DoesNotExist:
-                        return
-
-                # get answers from form
-                prefix='question_' #note this is hard coded in forms as well
-                for q in matching_questions:
-                    ans = form.cleaned_data['%s%d'%(prefix,q.pk)]
-                    if ans:
-                        create_or_update_response(q, ans)
-                    else:
-                        delete_response_if_exists(q, ans)
-
-
-                # Contract agreement
-                def create_signup():
-                    signup = None
-                    try:
-                        signup = SignupLog.objects.get(contact=contact, contract=contract, company = contact.belongs_to)
-                    except SignupLog.DoesNotExist:
-                        signup = SignupLog.objects.create(contact=contact, contract=contract, company = contact.belongs_to, type = 'complete')
-
-                # set exhibitor status to in progres if not already submitted
-                if exhibitor.status != 'complete_registration_submit' and exhibitor.status != 'complete_registration':
-                    exhibitor.status = 'complete_registration_start'
-                    exhibitor.save()
-
-                if form.accepting_terms():
-                    create_signup()
-
-                # Everything is done!
-                # Do nothing if form is saved, otherwise redirect and send email
-                save_or_submit = form.save_or_submit()
-                if 'submit' in save_or_submit:
-                    r.post(settings.SALES_HOOK_URL,
-                        data=json.dumps({'text': 'User {!s} just submitted complete registration for {!s}!'.format(contact, company)}))
-
-                    # log
-                    log = OrderLog.objects.create(contact=contact, company = contact.belongs_to, action='submit', fair=Fair.objects.get(current=True), products=productLog)
-                    log.save()
-
-                    # send email
-                    site_name = get_current_site(request).domain
-                    send_mail(
-                        'Complete Registration Confirmation on ' + site_name,
-                        get_template('register/complete_confirm_email.html').render(({
-                                'username': contact.email,
-                                'site_name': site_name,
-                                'bool_products': bool_products,
-								'num_products': num_products,
-                                'amount_len':max_amount,
-                                'name_len': max_name_len,
-								'base_price': BASE_PRICE,
-                                'total_price': total_price
-                            })
-                        ),
-                        settings.DEFAULT_FROM_EMAIL,
-                        [contact.email],
-                        fail_silently=False)
-
-                    # set exhibitor status to CR - submitted
-                    if exhibitor.status != 'complete_registration':
-                        exhibitor.status = 'complete_registration_submit'
-                        exhibitor.save()
-
-                    return redirect('anmalan:cr_done')
-                else:
-                    # create OrderLog
-                    log = OrderLog.objects.create(contact=contact, company = contact.belongs_to, action='save', fair=Fair.objects.get(current=True), products=productLog)
-                    log.save()
+                # a huge amount of stuff happens here, check help/ExhibitorForm.py for details
+                help.save_exhibitor_form(request, form, currentFair, company, contact)
 
     return render(request, template_name, {'form': form, 'contract_url': contract.contract.url})
+
 
 # thank you screen after submission of complete registration
 def cr_done(request, template_name='register/finished_registration.html'):
     return render(request, template_name)
+
 
 #change password
 def change_password(request, template_name='register/change_password.html'):
